@@ -29,7 +29,7 @@
   },
   "minimum-stability": "stable",
   "require": {
-    "php": "^7.1",
+    "php": ">=7.4",
     "ext-mbstring": "*",
     "ext-openssl": "*",
     "phpseclib/phpseclib": "^3.0",
