@@ -10,16 +10,11 @@ class JsonWebKeySetCollection
     /**
      * @var JsonWebKeySet[]
      */
-    private $authKeys;
+    private array $authKeys;
 
     public function __construct(array $authKeys)
     {
         $this->authKeys = $authKeys;
-    }
-
-    public function getAuthKeys(): array
-    {
-        return $this->authKeys;
     }
 
     public function getByCryptographicAlgorithm(string $algorithm): ?JsonWebKeySet
